backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
pandas==2.2.1
scikit-learn==1.4.2
